pyproject.toml
[tool.poetry]
name = "faktotum"
version = "1.0.0"
description = "Extracting information from unstructured text."
authors = ["Severin Simmler <severin.simmler@stud-mail.uni-wuerzburg.de>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.7.*"
syntok = "^1.2.2"
numpy = "^1.16.1"
pandas = "^0.25.3"
strsimpy = "^0.1.3"
transformers = "^2.5.0"
tqdm = "^4.41.1"
torch = "^1.4.0"
seqeval = { version = "^0.0.12", optional = true }
sklearn-crfsuite = { version = "^0.3.6", optional = true }
joblib = { version = "^0.14.1", optional = true }
tensorboard = { version = "^2.1.0", optional = true }
statsmodels = { version = "^0.11.0", optional = true }
flair = { version = "^0.4.5", optional = true }
scikit-learn = { version = "^0.22.1", optional = true }
scipy = { version = "^1.4.1", optional = true }
gensim = { version = "^3.8.1", optional = true }

[tool.poetry.dev-dependencies]
jupyterlab = "^1.2.5"
black = "^19.10b0"
isort = "^4.3.21"

[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"
